Maj extension

+++ extension/simulmans/src/main/java/fr/simulmans/Coords.java
@@ -1,4 +1,4 @@
 package fr.simulmans;
 
-public record Coords(int x, int y){
+public record Coords(double x, double y){
 }
